Symmetry Contour Visualization Added
Primary update is the addition of `symmetryContour.m`, which is an interactive app displaying the potential energy distribution on the sphere. One day I'll make the input to that function and `shapeWithSymmetry.m` more user friendly

For some reason I also decided that it would be worthwhile to change how coordinate conversions work. Having the elevation angle for spherical coordinates run from -pi/2 to +pi/2 is definitely more elegant and often more convenient than the standard 0 to pi. In retrospect is was absolutely not worthwhile, being probably 10 times more of a hassle than expected, but at least I got to clean up some code along the way.